feat: add sync progress grpc (#3722)
Description
---
Add sync progress grpc.  We do have `GetSyncInfo`, which returns block sync information. This returns the current state as well and the state of headersync. Problem is, that even with this information you will encounter scenario where after block sync is again a header sync. So doing some kind of progress bar have to be done with this in mind.

How Has This Been Tested?
---
Manually.