fix(core)!: add missing consensus encoding length byte rangeproof & covenants (#3730)
Description
---

- Adds consensus encode impl for RangeProof and ComSig
- Adds length varint for covenant
- Move version as first encoded byte
- Implement zero alloc hashing for outputs and witness hash
- Limit covenant and tariscript byte size
- Migration to set length byte in existing wallet transactions
- Ensure that an RPC message can never be zero length.
- covenant serde implementation does not use consensus encoding

Motivation and Context
---
Rangeproof is not fixed size and so requires a length varint. 

It is important to ensure that a remote peer cannot cause unbounded allocation, so various length checks are implemented. 

The wallet previously only ever stored a 0-byte covenant which is now invalid encoding.

Potential fix for WriteZero, ensuring that all rpc messages are non-zero in length (rare, but possible).

How Has This Been Tested?
---
Existing tests updated, manually - mined blocks on dibbler