fix: fix cucumber test for standard recovery
A recent fix to a major bug in the key manager that meant branching wasn’t being used broke the `Multiple Wallet recovery from seed node` cucumber test. This was because the test was trying to recover coinbases outputs whose keys are from a different branch in the key manager.

This PR fixes the `Wallet recovery with connected base node staying online` cucumber test to test recovering conventional outputs and marks `Wallet recovery with connected base node staying online` and `As a client I want to be able to restore my ffi wallet from seed words`as broken. They will be addressed by the team soon.