Mining node configuration
These changes let you specify mining node parameters in the
`[mining_node]` configuration section.

Before this, settings would be overwritten from the `global` config
struct. This is not what we want, because it prevents general Tari app
configuration. A classic example is that the base node GRPC address is
overwritten from the `base_node.grpc_base_node_address` config variable.

This typically is fine on a local machine setup, but what if the base
node and miner are on different networks (or in different docker
containers)?

`grpc_base_node_address` will be `0.0.0.0:18149` to let the node listen
for GRPC connections from other networks. But the mining_node
`base_node_addr` must be `/dns4/base_node_url/tcp/18149`, and NOT
`0.0.0.0:18149` as it is currently forced into.

The overwriting also undoes the work of the `DefaultConfigLoader` trait
which elegantly handles sub-configs and negates the need to carry
everything in a `GlobalConfiguration` struct.