feat(wallet_ffi)!:  add base node connectivity callback to wallet ffi
This PR adds a new callback to the FFI interface that fires when there is a change in the wallet’s connectivity to the set base node.

This information was already available from the Wallet Connectivity Service via a watch channel and this PR just hooks up this watch channel to the Wallet FFI callback handler.

The unit test for the callback handler is updated to test the addition.

feat: add specific LibWallet error code for “Fee is greater than amount”

Surfacing a specific error code for the “Fee is greater than amount” Transaction Build error.