fix(core)!: fix potential panic for sidechain merkle root with incorrect length (#3788)
Description
---
Fixes potential panic if a side-chain Merkle root of incorrect length is accepted by the blockchain.

Motivation and Context
---
Using a fixed 32-byte array forces any implementor to produce or validate the correct number of bytes. Esp. now that const generics are available, for performance and correctness, all hashes in the system should be fixed to 32 bytes and not held in a dynamic heap allocated vector.

I've changed the merkle_root type in SideChainCheckpoint because this is a relatively minor change.

This is a chain storage breaking change. 

How Has This Been Tested?
---
Existing tests, manually