feat(validator-node): initial state sync implementation (partial) (#3826)
Description
---
- add OpLogs
- add comms RPC methods and client calls to sync validator node state
- add synchronising state and basic implementation to validator node state machine (currently always in sync due to incorrect Merkle root in initial checkpoint)
- fix: fetching checkpoint from base node returned incorrect field for asset public key
- add read only interface for state db

Motivation and Context
---
Validator node on a committee needs to sync their state after going offline.

How Has This Been Tested?
---
Manually, it didnt work because the merkle root of the initial checkpoint is incorrect. Sync component has been hardcoded to always say it is in sync for now.