feat: add persistence of transaction cancellation reason to wallet db
This PR adds the persistence of the reason a transaction was rejected or cancelled to the db. This is done by expanding the existing `cancelled` field on a CompletedTransaction too not just be a boolean but an integer encoded nullable Enum which will encode the reason for the cancellation. By expanding the information contained by this field we no longer need the `valid` field which was used to indicate when a Coinbase Transaction was abandoned and we just make that a cancellation reason. A diesel migration is provided to handle this change is db schema.

Some specific areas changed to handle the update:
1. The console wallet was updated to take this change into account when displaying transactions.
2. The transaction validation protocol was updated to handle coinbases using this field rather than the valid field
3. Transaction database methods were updated to take the change into account

In the FFI we do the following:
1. Remove the `completed_transaction_is_valid` method for the valid that is taken out.
2. Added the `completed_transaction_get_cancellation_reason` method that returns whether the cancelled field is null as -1 or else the integer encoding of the TxCancellationReason enum.

```C
/// Gets the reason a TariCompletedTransaction is cancelled, if it is indeed cancelled
///
/// ## Arguments
/// `tx` - The TariCompletedTransaction
/// `error_out` - Pointer to an int which will be modified to an error code should one occur, may not be null. Functions
/// as an out parameter.
///
/// ## Returns
/// `c_int` - Returns the reason for cancellation which corresponds to:
/// | Value | Interpretation    |
/// |---  |---                  |
/// |  -1 | Not Cancelled       |
/// |   0 | Unknown             |
/// |   1 | UserCancelled       |
/// |   2 | Timeout             |
/// |   3 | DoubleSpend         |
/// |   4 | Orphan              |
/// |   5 | TimeLocked          |
/// |   6 | InvalidTransaction  |
/// |   7 | AbandonedCoinbase   |
/// # Safety
/// None
int completed_transaction_get_cancellation_reason(struct TariCompletedTransaction *transaction, int *error_out);
```

Tests are updated to take this change into account.
The Cucumber FFI test for cancellation is updated to test the new method