fix: improve sha3 pool mining (#3846)
Description
---
The current sha3 pool mining is mining at about 10% of the performance of the solo sha3 mining. This PR fixes the sha3 pool mining to use the same miner as the solo miner.  The solo and pool mining now gives the very same mining speed. 

Also changes the default logs to by default more clearly display what the miner is doing and how it performs. 

How Has This Been Tested?
---
Manual comparison.