feat: add logging of cancelled outputs when transaction is rejected
This PR just adds some logging of which outputs were specifically cancelled when a transaction is reject by the base node. In addition when an output is marked as Unspent we also clear the Mined In Block field which will result in that output being revalidated the next time a validation is done.