fix: fix flakey `test_coinbase_abandoned` integration test
This test has been flakey on CI and I think it is because when running release mode it is fast enough that the next stage in the test start before the last validation protocol in complete because it only waited for the cancellation events. This PR updates the test to wait for the validation completed event too to be sure that the previous validation protocol is complete before starting the next stage.