feat: add liveness call to wallet GRPC (#3854)
Description
---
Add GRPC call to check wallet liveness. The liveness service is mapped to GRPC now as well. 

Motivation and Context
---
We only have an FFI call to check the liveness but we need a GRPC option as well. This will allow any service connecting via GRPC to also call the liveness service to determine the liveness of the wallet.