feat!: add support for specifying custom messages for scanned outputs in lib wallet
This PR adds the ability for clients to specify custom text messages to be added to outputs that are recovered or one-sided payments that are scanned from the blockchain.
These messages were hardcoded before by clients should be able to specify their own so that localized messages can be provided in the future. The PR adds the ability to specify these messages when constucting the Utxo Scanner Service and also provides the interface to update these messages while the wallet is running.

In the FFI interface the `wallet_start_recovery` method was updated to accept a custom Recovery Output message. For the One-Sided payment custom message the `wallet_set_one_sided_payment_message` method was added to allow the client to set the One sided payment message after the Wallet has been created.

While working on this PR it was noted that the testing harness for the Utxo Scanner Service needed some updating and needed tests added that tested the the One-sided payment scanning. While writing these tests a bug was found in the Utxo Scanning logic that meant the tip clock was only scanned every second block that was received. This bug was fixed.