feat: update console wallet notifications
This PR creates console wallet notifications for transaction events and also adds a hotkey to clear notifications.

A bug was also spotted in the way that Faux transaction were validated that resulted in the last Faux transaction confirmed event to be emitted after each validation. The validation was updated to only emit this event if the transaction was previously not confirmed.