fix: add bound for number of console_wallet notifications
After sleeping on it I realised that notifications in the console wallet have no bound and an unattended wallet will fill its memory with notifications.

The notifications are ephemeral and do not have scrolling so this PR keeps just enough to be trivial in memory but fill a decent sized screen.