feat!: add recovery byte to output features (#3727)
Description
---
Added `recovery_byte` to `OutputFeatures`. The recovery byte will uniformly distribute all outputs in the blockchain into 255 "bins" that can be identified by the commitment of the output in combination with a master recovery byte key.

This change will come into effect on `dibbler` testnet once the `OutputFeatures` version is updated from `V0` to `V1`, which will result in a hard fork.

All funds will be retained after the hard fork.

Motivation and Context
---
The reasoning behind this is that it will enable a client to request only a specific subset of full-on outputs in a set of blocks. The commitment only takes up a fraction (less than 5%) of the byte space for a full-on output. A typical use case for this feature would be a wallet recovery client that needs to scan all outputs in the blockchain from the birthdate of the wallet to be recovered. The client would request all commitments in the set of blocks from a base node and identify all those that belong to their own "bin", after which a tailor made request can be submitted to retrieve the full-on output data for the identified commitments. This will result in a better than 20x saving in the amount of data transmitted for wallet recovery.

How Has This Been Tested?
---
- Unit tests
- Cucumber integration tests
- Performed system-level tests using a wallet that contains embedded output features in json text that does not contain the recovery byte.