fix: fix handling of creating faux transaction for recovered outputs
There existed a bug in the UTXO scanner that if a block had multiple outputs in it it would assign a single TxId for all the output but try import a separate faux transaction for each output. After the first output the rest would be rejected as duplicated outputs.

This PR updates the OMS scanning process to produce a unique TxId per output to avoid this issue.