chore: Change error propagation in validator node (#3811)
Description
---
This PR changes all (that can fail) `unwraps` to `?`.

Motivation and Context
---
There were too many unwraps that did not propagate the errors. It's better if we want to handle non fatal errors in the code instead of panicking .

How Has This Been Tested?
---
Just by running the validator node and collectibles.