[wip]: Implement qr code for base node -> mobile
Description

Create a QR deep link code pointing to `/base_nodes/add` according to the RFC-0154, so that the wallet can scan it and set the base node.

Motivation and Context
Users should be able easy to add a base node just scanning the QR printed in base node console.

How Has This Been Tested?
Manually, running the `tari_base_node` and scanning the QR using the camera and opening the URL with the Aurora app.
The base node should should be listed.