feat: react_based launchpad frontend
This is a long-lived feature branch. Do not merge it until complete.

This commit disables many of the long-running CI tasks that are not
relevant to this branch. HOWEVER, check the TODOs in the .github folder
when this branch gets merged into `development` and remove them, OR
cherry-pick to revery this specific commit.