test(cucumber): use separate FFI target dir (#4067)
Description
---
Uses `./temp/ffi-target` as the cargo target dir for FFI compilation.

Motivation and Context
---
The wallet and wallet FFI need to be recompiled for each run of the integration tests. This is because the compilation target differs between native and ffi, so previous compilations are overwritten. This PR sets a separate the target dir for FFI so that subsequent runs do not need to recompile the wallet and wallet FFI on each run.

How Has This Been Tested?
---
Locally, the wallet and FFI do not need to be recompiled after the initial run.