chore: update launchpad backend (#4017)
* feat: react_based launchpad frontend

This is a long-lived feature branch. Do not merge it until complete.

This commit disables many of the long-running CI tasks that are not
relevant to this branch. HOWEVER, check the TODOs in the .github folder
when this branch gets merged into `development` and remove them, OR
cherry-pick to revery this specific commit.

* Remove deprecated useBootstrapper option

`useBootstrapper` is no longer a valid configuration option in
tauri.conf.json, so we remove it.

* Switch to public minideb base package

The quay.io version of bitnami's minideb package does not seem to work
anymore.
The commit also leverages `install_packages` which removes the
boilerplate of having to `apt update && apt install .. & rm
/var/dpkg/...`

Add platform (arm64 / amd64) to docker tag and switch minideb source

Add the platform marker to all docker image tags

* Add architecture-dependent image resolution

Add a function to use the appropriate docker image based on CPU architecture, e.g. latest-arm64 for M1 chips, or latest-amd64 for Intel chips.

* Configure static base node parameters in config file

With the new config setup, some config setting names are no longer
network-dependent, so they can be moved into the lauchpad config.toml
template.

* Update docker configs for wallet and mm_proxy

Moves static config variables to the config.toml file, and dynamic
variables are updated according to the config variable names.

* Pull out wallet command

Keep interface consistent with rest of code and provide a `wallet_cmd`
function for the docker execution command arguments.

Fixes a small envar string typo.

* Update SHA3 miner image config

* Update config.toml [miner] section so that it connects to the wallet
  and base node
* Pull command line arguments into its own procedure, in line with other
  images

* Add multiple monero stagenet urls as default

* fix: issues with launchpad backend  (#4074)

Description
--

Fixes issues with launchpad backend. See individual commit messages for changes.

How Has This Been Tested?
---
Manually