wip: Decoupling backend and UI. Implementing new tour command: network_list and restart all containers.
Description

Splitting the Tauri API contract from the underlying backend services.

Implement network_list service and expose it to Tauri command.

Implement `restart all services` functionality which will be called on network change event:
	- stop all containers with status Running
	- start the already stopped containers with new network configuration
	- TODO: execution order should be implemented as there is a strict staring sequence.
The restart all containers functionality will be applied when the user changes the network.

Decoupling the workspace and docker services.
	- start and stop container calls are changed from struct impl. to functions

Implementing docker compose start, stop and ps commands.
Gather statistic and logs from running containers.

Motivation and Context

At the moment the `workspace` is responsible for configuring user's workspace and run the docker containers on top of it.
As those are 2 independent actions it is better to split them.
Workspace will be responsible for user's configuration whether the docker interaction will be moved to independent functions.
Another reason for the change is that workspace is coupled with `Tauri` app whether managing docker containers (create, stop, start, etc.) should be independent from the running context.
As side effect the testing is a real challenge.

How Has This Been Tested?
Launchpad is started and all docker images are started and stopped.
Logging, statistic work.
Started and stopped containers successfully.