fix(wallet): ensure block hash exists (#4083)
Description
---
Adds a quick data validation check before updating the database. The base node may return a response without a block_hash and the resulting db update creates an invalidatable scenario for the wallet.

Motivation and Context
---
An aurora wallet is failing to validate and causing problems

How Has This Been Tested?
---
It has not yet been tested.