fix(core): set reorg metrics on header sync rewind (#4087)
Description
---
Sets `tari_base_node::blockchain::reorgs` and `tari_base_node::blockchain::tip_height` metrics on header sync rewind.

Motivation and Context
---
Header sync is designed to handle syncing from nodes that are on a forked chain, and so can rewind and reorg the blockchain. This PR sets the correct metrics for a header sync rewind.

How Has This Been Tested?
---
Not explicitly tested in this PR (reorgs tested in general and in cucumber tests). This case is fairly rare and difficult to test manually (unit tests should be implemented for header sync! but this is out of scope). The correct metrics should be seen on grafana over the next month.