fix(ci): sort .license.ignore locally before diff (#4106)
Description
---

- sorts license.ignore localy before diffing

Motivation and Context
---
The license check does not always pass. I'm not really able to explain why it is flaky but on my local linux machine
it does not pass because the `sort` command behaves slightly differently. 

This PR sorts the ignore file locally so that we can be sure the sorting is consistent before the diff.

How Has This Been Tested?
---
Manually run ./scripts/file_license_check.sh