feat: contract auto acceptance (#4177)
Description
---
Add the configuration for VN auto acceptance. After finding constitutions you can be a part of, auto respond with an acceptance.

Motivation and Context
---
to make it very easy for VN to become part of committees.


How Has This Been Tested?
---
~YOLO.~
~I joke, i'll write a spec for it before merge.~

Spec written but we've got some other bits we need to work out to get the specs passing again.

Tested manually and the process worked but transaction is rejected by the base node due to an input bug trying to utilize the contract as an input.