fix(wallet): use correct type for contract_id in the contract constitution file format (#4179)
Description
---
Changed the type of the `contract_id` field in the file format from a `PublicKey` to a `String`, and then parsed it as hex into a `FixedHash` (which is the proper type).

Motivation and Context
---
When using a real `contract_id` value on contract constitution publishing, we got a file format parsing error

How Has This Been Tested?
---
Changed the `contract_id` field in the fixture file to a value that trigger the issue. The test now pass.