feat(validator_node): add global db (#4210)
Description
---
This PR intorduces a new global database for the validator node. This is so it can store relevant state to help process contracts such as:
- the constitutions it's a part of so it can query acceptances
- the last scanned block hash so we don't always scan from genesis

Motivation and Context
---
Everytime the VN would scan for constitutions it would do so from the genesis block. It would then re-accept the same constitutions on every loop. We had no way to keep track of already scanned block, or what constitutions we're a part of. Long term we needed a solution and not a bandaid so we introduced a global state db.

How Has This Been Tested?
---
Manually. Ensuring the VN no longer eager accepts constitutions non-stop. (not actually tested yet).

* feat(blockchain-db): add {block_hash, output_type, output_hash} index to contract_index

* docs(chain_storage/lmdb-db): document mapping for each lmdb database

* feat(chain-storage): replace fetch_all_constitutions with fetch_contract_outputs_for_block

* test(chain-storage): add unit tests for fetch_contract_outputs_for_block

* feat(base-node-service): replace get_constitutions with get_contract_outputs_for_block

* feat(base-node-grpc): use get_contract_outputs_for_block to fetch constitutions

* test(cucumber): mark `Publish contract acceptance` as broken

* test(cucumber):           fix command to publish contract utxos

* fix(chain-storage): fetch_chain_headers should return empty vec if query is out of range

* Introduce the VN global db

This db is for the validator node to store global state information and
data it needs to process information. For example:
 - the constitutions it's a part of so it can query acceptances
 - the last scanned block hash so we don't always scan from genesis

* Intro the db to the VN

* Fix default relative path

* Use the latest block hash when scanning

* Refactor

* Update the defaults

* Ensure we commit the transaction to save the last hash

* Dont include the last scanned block in the next batch

Co-authored-by: Stanimal <sdbondi@users.noreply.github.com>