feat: change tari explorer block view (#4226)
Description
---
Make partials (this is the name for handlebars component) for each proto message (e.g. message ConstitutionChangeRules is in ConstitutionChangeRules.hbs). This way it's more readable code, and easily extensible.
The fields that are empty are not in the output.

Motivation and Context
---
Reading json for the features was not ideal way.
Now everything is in readable format.
Example output:
![image](https://user-images.githubusercontent.com/35243812/175246425-bcc490bf-bf4e-431c-a90e-0dc0fb19e8fc.png)


How Has This Been Tested?
---
Manually. But I haven't seen all output types. If we add new output_type, it will be shown as "Standard" due to how lookup works. This should be kept in mind.