feat(vn): record contract states (#4241)
Description
---
Add a contracts table to the global vn db so we can store the state of contracts the vn is a member of.

Motivation and Context
---
It will make it easier to go back and check known contracts that are awaiting quorum acceptance.

How Has This Been Tested?
---
Manually, and by fixing the previously broken spec.

Commits
---

* Add the constitutions table to the globaldb

* Add contract saving call

* Refactor the naming from constitution to contract

* Save the contract in a pending state

* Store expired contracts

* Update contract states

* Only auto accept if the setting is set

* Make the sleep length configurable

The long sleep length was causing the specs to fail. We also don't want
to wait in minutes for polling to occur. Make the option configurable
and use a much shorter time in testing.

* Rename to state for consistency

* Fix clippy warnings

* No more lockfiles, npm install it